chore: Downgrade workmanager to v0.8.0
This commit downgrades the `workmanager` package from version `0.9.0+3` to `0.8.0`.

This change was necessary to resolve an issue in the `BackgroundTaskService`, where the `networkType` constraint parameter was updated. The newer version of the package uses `NetworkType.notRequired`, while the older version expects `NetworkType.not_required`.

The dependency has been updated in `pubspec.yaml` and `pubspec.lock` to ensure compatibility.